requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
numpy==2.1.1
pillow==10.4.0
reportlab==4.2.2
sqlparse==0.5.1
typing_extensions==4.12.2
